Refactor external pullers to separate files
